Class #131: Unit Testing - Intro 2

+++ CSharp/_12_UnitTesting/_02_MyClass.cs
@@ -4,6 +4,6 @@ public class MyClass
 {
     public static decimal division(int dividend, int divisor)
     {
-        return dividend / divisor;
+        return dividend / (decimal)divisor;
     }
 }
